# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Parali Se Paisa Action Platform API",
    "description": "Offline-first prototype API. Demo records remain marked synthetic; this is not an operational government service.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/public/certificates/{certificate_id}": {
      "get": {
        "tags": [
          "public certificate verification"
        ],
        "summary": "Public Verify",
        "operationId": "public_verify_api_v1_public_certificates__certificate_id__get",
        "parameters": [
          {
            "name": "certificate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Certificate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/certificates/{certificate_id}/file": {
      "get": {
        "tags": [
          "public certificate verification"
        ],
        "summary": "Public Pdf",
        "operationId": "public_pdf_api_v1_public_certificates__certificate_id__file_get",
        "parameters": [
          {
            "name": "certificate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Certificate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/balers": {
      "get": {
        "tags": [
          "balers"
        ],
        "summary": "List Balers",
        "operationId": "list_balers_api_v1_balers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "balers"
        ],
        "summary": "Create Baler",
        "operationId": "create_baler_api_v1_balers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/balers/{baler_id}": {
      "patch": {
        "tags": [
          "balers"
        ],
        "summary": "Update Baler",
        "operationId": "update_baler_api_v1_balers__baler_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "baler_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Baler Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BalerPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/buyers": {
      "get": {
        "tags": [
          "buyers"
        ],
        "summary": "List Buyers",
        "operationId": "list_buyers_api_v1_buyers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "buyers"
        ],
        "summary": "Create Buyer",
        "operationId": "create_buyer_api_v1_buyers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/buyers/{buyer_id}": {
      "patch": {
        "tags": [
          "buyers"
        ],
        "summary": "Update Buyer",
        "operationId": "update_buyer_api_v1_buyers__buyer_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "buyer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Buyer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyerPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/certificates": {
      "get": {
        "tags": [
          "certificates"
        ],
        "summary": "Certificates List",
        "operationId": "certificates_list_api_v1_certificates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/certificates/{certificate_id}": {
      "get": {
        "tags": [
          "certificates"
        ],
        "summary": "Certificate Get",
        "operationId": "certificate_get_api_v1_certificates__certificate_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "certificate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Certificate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/certificates/{certificate_id}/file": {
      "get": {
        "tags": [
          "certificates"
        ],
        "summary": "Certificate File",
        "operationId": "certificate_file_api_v1_certificates__certificate_id__file_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "certificate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Certificate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/certificates/generate/{verification_id}": {
      "post": {
        "tags": [
          "certificates"
        ],
        "summary": "Certificate Generate",
        "operationId": "certificate_generate_api_v1_certificates_generate__verification_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "verification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Verification Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BalerCreate": {
        "properties": {
          "baler_id": {
            "type": "string",
            "title": "Baler Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "operator_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Name"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "current_location_lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Lat"
          },
          "current_location_lon": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Lon"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "BUSY",
              "OFFLINE",
              "MAINTENANCE"
            ],
            "title": "Status",
            "default": "AVAILABLE"
          },
          "working_start": {
            "type": "string",
            "title": "Working Start",
            "default": "08:00"
          },
          "working_end": {
            "type": "string",
            "title": "Working End",
            "default": "17:00"
          },
          "service_rate_acres_per_hour": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Service Rate Acres Per Hour"
          },
          "daily_capacity_acres": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Daily Capacity Acres"
          },
          "max_straw_capacity_tonnes": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Straw Capacity Tonnes"
          },
          "available_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available From"
          },
          "demo_or_real": {
            "type": "string",
            "enum": [
              "DEMO",
              "SYNTHETIC",
              "FIXTURE",
              "REAL"
            ],
            "title": "Demo Or Real",
            "default": "DEMO"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "API_UNVERIFIED_INPUT"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "baler_id",
          "name",
          "latitude",
          "longitude",
          "service_rate_acres_per_hour",
          "daily_capacity_acres"
        ],
        "title": "BalerCreate"
      },
      "BalerPatch": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "operator_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operator Name"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "current_location_lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Lat"
          },
          "current_location_lon": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Location Lon"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "AVAILABLE",
                  "BUSY",
                  "OFFLINE",
                  "MAINTENANCE"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "working_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Start"
          },
          "working_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working End"
          },
          "service_rate_acres_per_hour": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Rate Acres Per Hour"
          },
          "daily_capacity_acres": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Capacity Acres"
          },
          "max_straw_capacity_tonnes": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Straw Capacity Tonnes"
          },
          "available_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available From"
          },
          "demo_or_real": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "DEMO",
                  "SYNTHETIC",
                  "FIXTURE",
                  "REAL"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Or Real"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "BalerPatch"
      },
      "BuyerCreate": {
        "properties": {
          "buyer_id": {
            "type": "string",
            "title": "Buyer Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "buyer_type": {
            "type": "string",
            "enum": [
              "BIOMASS_POWER",
              "CBG",
              "PELLET",
              "INDUSTRIAL_BOILER",
              "OTHER"
            ],
            "title": "Buyer Type",
            "default": "OTHER"
          },
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "accepted_residue_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Accepted Residue Types"
          },
          "daily_demand_tonnes": {
            "type": "number",
            "minimum": 0.0,
            "title": "Daily Demand Tonnes"
          },
          "remaining_demand_tonnes": {
            "type": "number",
            "minimum": 0.0,
            "title": "Remaining Demand Tonnes"
          },
          "max_moisture_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Moisture Pct"
          },
          "price_per_tonne": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Per Tonne"
          },
          "price_is_verified": {
            "type": "boolean",
            "title": "Price Is Verified",
            "default": false
          },
          "operating_start": {
            "type": "string",
            "title": "Operating Start",
            "default": "00:00"
          },
          "operating_end": {
            "type": "string",
            "title": "Operating End",
            "default": "23:59"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE",
              "PAUSED"
            ],
            "title": "Status",
            "default": "AVAILABLE"
          },
          "demo_or_real": {
            "type": "string",
            "enum": [
              "DEMO",
              "SYNTHETIC",
              "FIXTURE",
              "REAL"
            ],
            "title": "Demo Or Real",
            "default": "DEMO"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "API_UNVERIFIED_INPUT"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "buyer_id",
          "name",
          "latitude",
          "longitude",
          "daily_demand_tonnes",
          "remaining_demand_tonnes"
        ],
        "title": "BuyerCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}